LayoutTests/fast/notifications/notifications-replace.html

<!DOCTYPE html>
<html>
<head>
    <script type="text/javascript">
        function log(message)
        {
            document.getElementById("result").innerHTML += message + "<br>";
        }
        
        function runTests()
        {
            if (window.layoutTestController) {
                layoutTestController.grantDesktopNotificationPermission("file://");
                layoutTestController.dumpAsText();
            }

            if (!window.webkitNotifications) {
                log("FAIL: No webkitNotifications interface!");
            }
            
            var M = window.webkitNotifications.createNotification("", "New E-mail", "Meet me tonight at 8!");
            M.replaceId = "NewMail";
            M.show();
            var other = window.webkitNotifications.createNotification("", "Alarm", "It's time to wake up!");
            other.show();
            var N = window.webkitNotifications.createNotification("", "New reply", "Correction, 8:15");
            N.replaceId = "NewMail";
            N.show();          
        }
    </script>
</head>
<body>
<p>Showing notifications with replace ID...</p>
<div id="result"></div>
<script type="text/javascript">
runTests();
</script>
</body>
</html>
