safari-600.7-branch/LayoutTests/fast/notifications/notifications-replace-expected.txt

DESKTOP NOTIFICATION: icon , title New E-mail, text Meet me tonight at 8!
DESKTOP NOTIFICATION: icon , title Alarm, text It's time to wake up!
REPLACING NOTIFICATION New E-mail
DESKTOP NOTIFICATION: icon , title New reply, text Correction, 8:15
Showing notifications with replace ID...
